nios2: IIC Controller tweaks
This patch follows the Nios2 IIC controller specification
a bit more closely, in particular in emulating the control
register "ipending".

Added two new callback routines:
    nios2_iic_update_cr_status
    nios2_iic_update_cr_ienable
These routines need to be called whenever the "status" or "ienable"
control registers are modified in order to re-evaluate interrupts.

Added the routine "nios2_iic_create", this simplifies adding
the IIC controller into various devices as the common code
does not need to be replicated in each device that uses the IIC.

Signed-off-by: Juro Bystricky <juro.bystricky@intel.com>